test: fix test cases with minor refactoring
