+++ .gitignore
@@ -1,1 +1,4 @@
 *.test
+.dapper
+coverage.out
+
