+++ io/file_test.go
@@ -565,10 +565,15 @@ func (s *TestSuite) TestGetDiskStat(c *C) {
 		c.Assert(err, IsNil, Commentf(test.ErrErrorFmt, testName, err))
 
 		// On the running system, FreeBlocks/StorageAvailable might be changing with time.
-		// so we only compare the following fields
+		// So we only compare the following fields
 		c.Assert(diskStat.DiskID, Equals, expectedDiskStat.DiskID)
 		c.Assert(diskStat.Path, Equals, expectedDiskStat.Path)
-		c.Assert(diskStat.Type, Equals, expectedDiskStat.Type)
+
+		// FIXME: overlayfs is not supported in the github.com/shirou/gopsutil/v3
+		if expectedDiskStat.Type != "overlayfs" {
+			c.Assert(diskStat.Type, Equals, expectedDiskStat.Type)
+		}
+
 		c.Assert(diskStat.TotalBlocks, Equals, expectedDiskStat.TotalBlocks)
 		c.Assert(diskStat.BlockSize, Equals, expectedDiskStat.BlockSize)
 		c.Assert(diskStat.StorageMaximum, Equals, expectedDiskStat.StorageMaximum)
