+++ net/net_test.go
@@ -1,0 +1,15 @@
+package net
+
+import (
+	"testing"
+
+	. "gopkg.in/check.v1"
+
+	_ "github.com/longhorn/go-common-libs/test"
+)
+
+func Test(t *testing.T) { TestingT(t) }
+
+type TestSuite struct{}
+
+var _ = Suite(&TestSuite{})
